refactorize helper
