+++ src/dev/task/buildWiki.ts
@@ -1,7 +1,6 @@
 import fs from 'fs'
 import path from 'path'
-import { Helper } from '../../lib/manager/helper'
-import { orm } from '../../lib'
+import { orm, Helper } from '../../lib'
 
 async function writeFunctions (category:string, list: any): Promise<void> {
 	const lines: string[] = []
