+++ src/lib/index.ts
@@ -1,6 +1,6 @@
 import { Orm } from './orm'
 export * from './model'
 export * from './repository'
-export { Helper } from './manager/helper'
+export { Helper } from './manager'
 export { Orm } from './orm'
 export const orm = Orm.instance
