+++ src/lib/manager/index.ts
@@ -1,3 +1,4 @@
+import { OrmHelper } from './helper'
 export * from './expressionManager'
 export * from './expressionNormalizer'
 export * from './transaction'
@@ -7,10 +8,10 @@ export * from './schema'
 export * from './stageFacade'
 export * from './ddlBuilder'
 export * from './routing'
-export * from './helper'
 export * from './operandManager'
 export * from './language'
 export * from './ddlBuilder'
 export * from './languages'
 export * from './dialect'
 export * from './sentenceCompleter'
+export const Helper = new OrmHelper()
