+++ src/lib/manager/stageFacade.ts
@@ -7,7 +7,7 @@ import { StageExport } from '../stage/stageExport'
 import { StageTruncate } from '../stage/stageTruncate'
 import { StageImport } from '../stage/stageImport'
 import { StageDelete } from '../stage/stageDelete'
-import { Helper } from './helper'
+import { Helper } from './'
 import { SchemaError, Stage, View, OrmOptions } from './../model'
 
 export class StageFacade {
